test(daemon/crash): move pruner/raw-appender/sources specs into src/__tests__ for unit coverage gate (Task #474)
The crash retention pruner, raw-appender, and capture-sources modules
already had thorough specs under `packages/daemon/test/crash/`, but those
files live outside the unit-coverage gate's include glob — the coverage
config (`vitest.config.coverage.ts`) only counts `src/**/*.spec.ts` and
`src/**/__tests__/**/*.spec.ts`. Result: src/crash measured 14.1% line
coverage on origin/working despite the modules being well-exercised by
integration-tier specs.

Move the four relevant spec files into `packages/daemon/src/crash/__tests__/`
and adjust the import paths from `../../src/crash/...` -> `../...` and
`../../src/db/...` -> `../../db/...`. No production code or test logic
changes — only the file locations and matching import paths.

Coverage delta on src/crash (per `vitest run --coverage --config vitest.config.coverage.ts`):

| File                    | before | after   |
| ----------------------- | ------ | ------- |
| src/crash (folder line) | 14.10% | 94.19%  |
| pruner.ts               |  0.00% | 96.87%  |
| raw-appender.ts         | 12.12% | 96.96%  |
| sources.ts              | 12.65% | 88.60%  |

All three target files exceed the 75% line-coverage bar. pruner-decider
is left at its existing test/crash location per task scope (Task #435 already
covered the filter-decider variant; pruner-decider is now indirectly covered
to ~100% via pruner.spec.ts importing pruner.ts).

Note: Task #435 already shipped the filter-decider portion of the broader
crash coverage push; this PR closes the remaining pruner/raw-appender/sources
gap and is the final piece needed before the unit-coverage threshold can move
toward the spec ch12 §6 ship target of 80%.
